Reorganize EnablePayments into Pay/Wallet/shared structure
