+++ src/pages/EnablePayments/Pay/Onfido/OnfidoPrivacy.tsx
