+++ src/pages/EnablePayments/Pay/Onfido/OnfidoStep.tsx
