+++ src/pages/EnablePayments/Wallet/FeesAndTerms/FeesAndTerms.tsx
