+++ src/pages/EnablePayments/Wallet/FeesAndTerms/substeps/TermsStep.tsx
