+++ src/pages/EnablePayments/Wallet/PersonalInfo/substeps/ConfirmationStep.tsx
@@ -4,7 +4,7 @@ import useLocalize from '@hooks/useLocalize';
 import useOnyx from '@hooks/useOnyx';
 import type {SubStepProps} from '@hooks/useSubStep/types';
 import {getLatestErrorMessage} from '@libs/ErrorUtils';
-import getSubstepValues from '@pages/EnablePayments/utils/getSubstepValues';
+import getSubstepValues from '@pages/EnablePayments/Wallet/utils/getSubstepValues';
 import CONST from '@src/CONST';
 import ONYXKEYS from '@src/ONYXKEYS';
 import INPUT_IDS from '@src/types/form/WalletAdditionalDetailsForm';
