+++ src/pages/EnablePayments/Wallet/VerifyIdentity/VerifyIdentity.tsx
