+++ src/pages/EnablePayments/shared/FailedKYC.tsx
