+++ src/pages/EnablePayments/shared/TermsForms/LongTermsForm.tsx
