+++ src/pages/EnablePayments/shared/TermsForms/ShortTermsForm.tsx
