+++ src/pages/EnablePayments/shared/useHasFreshWalletData.ts
